refactor: MonthGridBuilder.kt 파일을 제거하고 관련 로직을 MonthGrid.kt로 통합
- `MonthGridBuilder.kt` 파일을 삭제하고 `buildMonthGrid` 함수를 `MonthGrid.kt` 내부의 private 함수로 이동하였습니다.
- 별도의 파일로 관리되던 그리드 생성 로직을 컴포넌트 파일로 통합하여 응집도를 높였습니다.
